Reorganize pipeline result plotting into its own module
Trigger CI
