+++ tox.ini
@@ -120,6 +120,7 @@ commands =
         src/pykeen/utils.py \
         src/pykeen/datasets/ \
         src/pykeen/triples/utils.py \
+        src/pykeen/pipeline_plot.py \
         setup.py
 description = Run the flake8 tool with darglint.
 
@@ -155,6 +156,7 @@ commands = mypy --ignore-missing-imports \
     src/pykeen/constants.py \
     src/pykeen/losses.py \
     src/pykeen/optimizers.py \
+    src/pykeen/pipeline_plot.py \
     src/pykeen/regularizers.py \
     src/pykeen/typing.py \
     src/pykeen/utils.py \
